Initial xCelerator documentation.
This version is not yet complete, but it contains a basic structure of
what the documentation will look like. The resulting documentation will
be a PDF created by Apache FOP. The FOP templates will be added to the
source code later. Currently there is no PDF generated from this
documentation.