Added xml transform to generate pdf documentation.
This commit adds a dependency on apache FOP to be able to build the system.
It has to be configured in each environment the system is built with.

The documentation is written in a simple custom made xml format which can
express the most basic functionality that is needed to document this
xCelerator.